backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
python-jose[cryptography]>=3.3,<4
bcrypt>=4.1,<6
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.29,<1
psycopg[binary]>=3.1,<4
alembic>=1.13,<2
boto3>=1.34,<2
celery[redis]>=5.3,<6
redis>=5.0,<6
jinja2>=3.1,<4
weasyprint>=62,<70
python-multipart>=0.0.9,<1
httpx>=0.27,<1
pydantic>=2.6,<3
pydantic-settings>=2.1,<3
pytest>=8,<9
pytest-asyncio>=0.23,<1
